requirements.txt
Django==1.8.1
PyYAML==3.11
django-debug-toolbar==1.3.0
django-filters==0.1.0
django-rest-swagger==0.2.9
djangorestframework==3.1.1
sqlparse==0.1.15
wsgiref==0.1.2
